moving funcs from ptyrex.py to ptycho_utils.py - deleting the ptychography folder
